Verify that all coordinates passed to addcoords() are correct.
Check types, values, dimensions and valid sizes for x,y,Z coordinates.

One thing we don't check for is an empty {x,y,z}range, which still produces a
gnuplot warning. I think this is a case of a gnuplot warning we want to
preserve, but may revisit this in the future.

Also, fix a small bug with addcoords(X) -- should be addcoords(Y) instead.