Add validation and default configuration functions.
We want to add functions to change Gaston's defaults for figure and curve
properties. We would like these functions to only accept valid values. This
means creating a set of validation functions which, it turns out, may also
be used in other points of the program.

We currently support setting a default terminal type, a default legend and
a default plotstyle.