Add delay between tests.
When running tests, sometimes gnuplot would barf on the screen and complain
about missing or incomplete data from a file. The reason was that gnuplot_send
sends the plot commands but doesn't wait for gnuplot to finish. So, the tests
were running too fast for gnuplot: we would generate a new data file before
gnuplot had had a chance to read the previous one. The solution is to add a
small delay (currently 0.2 seconds) between tests. On slower machines, this
number may have to be increased (this should really be a settable value and
not hard-coded).