Preliminary support for printing to file.
Printint to PDF seems to work fine. The following terminal properties are
supported for printing:

mono/color, fontface, fontsize, fontscale, linewidth, and size.

The way printing is supported is a bit hackish but seems to work well: we
temporarily set the terminal type to a "file type", replot, and restore the
terminal type.