Make imagesc 50% faster
Before this change, we were writing each row of data to a file
using writedlm. This involves creation of a temporary array for
each line. Now, we pre-allocate an array, write the data to it, and
then save the entire array to disk.

On my machine, @time of imagesc of a 1000x1000 random matrix went
from 3.6 seconds to 1.3 seconds, and gc time went from 15% to 3%.

A quick profile view does not show any low-hanging fruit to optimize
this further.