Fix race condition when plotting consecutively to the same figure.
The race condition occurs when issuing two plot commands in a very short
time interval. Gnuplot may still be reading the data file for the first
plot when Gaston runs the second command and modifies the file.

This is fixed by having gnuplot print a message to STDOUT. Gaston's
`llplot()` waits until the message is ready in its pipe to gnuplot's
STDOUT. This prevents a subsequent plot command to run before the first
has completed.