Read and print any gnuplot output to STDERR from `llplot()` command.
Any gnuplot output to STDERR means that something went potentially
wrong. Before, we let STDERR spill into julia's REPL. Now, thanks to our
`popen3()` function, we capture it. After plotting, we check the STDERR
pipe; if it is not empty, we read it, print a warning, and print
gnuplot's message.