Make the testsuit conform to Julia's package standards.
Julia offers a test framework that makes our own macros unnecessary.
Move the test file to test/ and rename it to runtests.jl.

Some functions (e.g. plot) had to be modified to return a consistent
value (the figure handle) in all cases. The result is that Gaston
behaves more consistently now.