Reduce the number of temp files created.
Instead of creating one new random temporary file per figure, re-use
files with form `/tmp/gaston-(randstring)-handle`. The random part of
the filename is determined at startup time and remains the same for the
session. This allows several instances of Gaston to run at the same time
on the same filesystem.