Fix default size for saving to pdf. Fixes #74.
Some terminals (like svg and gif) want a size in pixels, and other
terminals (like pdf) want a size in inches. The default size was set
to "640,480" and the default print term is "pdf", resulting a huge
pdf with labels so small that they couldn't be read.

This changes the default size to "5in,3in", which is appropriate for
PDF-like terminals, and sets it to "640,480" if inside jupyter.

It is up to the user to select a correct size in pixels if they switch
to GIF-like terminals.