Rework fetching of gnuplot output
The new code is more robust, faster and simpler.

The old code could deadlock when gnuplot sent large SVGs to stdout, as
was happening at least in Jupyter notebooks. The deadlock occurred
because the ChanStdOut had limited capacity. When both the channel and
OS kernel pipe buffer were full, gnuplot blocked waiting for Gaston to
consume the output. But this never happened, because Gaston was
waiting for "GastonDone" on stderr, which gnuplot should have sent
after sending the whole figure. But gnuplot was blocked...

Now, gnuplot's output is consumed all the time during plotting.

We also try to detect better the start and end of the sent data by
inserting markers to both stdout and stderr streams and only process
the data between those markers.

The new code is faster, because it does not use sleeps to wait for
gnuplot. If gnuplot is able to supply the data quickly, we consume
them immediately and finish just when we see the end marker, without
additional delays.

On the other hand, when gnuplot is slow, we just wait in
readavailable(). We use timeout timer to handle the case when the end
marker does not arrive for whatever reason in reasonable time.

Fixes #103