Allow specifying size of dumb terminals
This is useful one wants to see ASCII art graphs in the terminal and
is not satisfied with the default size.

Unfortunately, this commit makes dumb terminal not consistent with
other terminals, where print_size is used only for printing to file.
With current implementation, when "dumb" is added to term_file, it is
no longer possible to see ASCII art in the terminal.