HUGE refactoring of configuration code
I want Gaston to support setting terminal options on a plot-by-
plot basis; for example, plot with a blue background one time and
then revert to the default. However, gnuplot doesn't support
reverting a terminal to its default state: once an option (such as
background) is set, the default is lost forever. Supporting this use
case, then, means that Gaston must know gnuplot's defaults for all
supported terminals, and must be able to reset them. This prompted a
review of all the option-setting code.