Allow user to set timeout values
It is well known that the main difficulty of using Gnuplot is that
we need to interact with it via the standard pipes.

It turns out that different systems have different pipe speeds, and
sometimes gnuplot just takes too long to initialize, and the OS is
too slow to send data back and forth.

With this change, users can set a timeout that is appropriate for
their system.

Use

    `set(stdout_timeout = X)`

to set the `stdout` timeout to `X` seconds, and

    `set(stderr_timeout = X)`

to set the `stderr` timeout to `X` seconds.

In Unix-like systems, the default is 1 and 5 seconds; in Windows, it is
20 seconds for both.

Fixes #118