Internal type changes
Changed the way the long argument lists for `plot` and `plot!`
are handled. Now the arguments are taken from `args...` and
are stored in a struct. The aim is clearer code and more
flexibility when adding arguments.

This should pave the way to enabling the use of symbols and numbers for
plot configuration (in addition to strings).