Simplify Gaston/Gnuplot interaction
Now that plot data (such as svg and png) is passed through files,
it is no longer necessary to store this data in Figure.svg. Also,
it is not necessary to read it from Gnuplot's stdout. We still wait
for Gnuplot to reply to a `print Gastondone` command, to make sure
that Julia is not blocked permanently if Gnuplot fails. We also
check Gnuplot's stderr to catch any errors.