Massive rewrite (towards 1.0!)
In this rewrite, I have removed most of the limitations of Gaston when
interacting with Gnuplot.

In past versions, I was too concerned about never doing anything that
might produce an error in gnuplot, and that I hadn't personally verified
to work correctly. This approach is very safe but limits what Gaston can
ask gnuplot to do (which turned out to be very little).

This new version lifts all these limitations, and relies on gnuplot to
catch and report errors. As of now, _most_ of what gnuplot can do should
be possible from Gaston.

A few limitations remain:
    * To support the Matlabish "figure" commands, I need to break up
      the termstring into `term` and `termopts` sections.
    * I still require all data to be numeric (label plots are not
      supported _yet_)
    * Gnuplot is used almost exclusively in 'datafile' mode.
    * Probably a few others I haven't found out about yet.

After some more testing, debugging and documentation, this should become
Gaston 1.0 (after more than 8 years!)